+++ WebApp/public/bidirectional/index.html
@@ -26,7 +26,7 @@ title="View source for this page on GitHub" id="viewSource">View source on GitHu
 <script src="https://unpkg.com/event-target@latest/min.js"></script>
 <script src="https://unpkg.com/resize-observer-polyfill@1.5.0/dist/ResizeObserver.global.js"></script>
 <script src="https://cdn.polyfill.io/v2/polyfill.min.js?features=IntersectionObserver"></script>
-<script type="module" src="js/app2.js"></script>
+<script type="module" src="js/main.js"></script>
 
 </body>
 </html>
