+++ WebApp/public/bidirectional/js/main.js
