+++ WebApp/public/js/register-events.js
