+++ WebApp/public/videoplayer/index.html
@@ -2,12 +2,12 @@
 <html>
 <head>
   <link rel="icon" href="../images/favicon.ico" type="image/x-icon">
-  <link type="text/css" rel="stylesheet" href="stylesheets/style.css">
+  <link type="text/css" rel="stylesheet" href="css/style.css">
   <script type="text/javascript" src="https://webrtc.github.io/adapter/adapter-latest.js"></script>
   <script src="https://unpkg.com/event-target@latest/min.js"></script>
   <script src="https://unpkg.com/resize-observer-polyfill@1.5.0/dist/ResizeObserver.global.js"></script>
   <script src="https://cdn.polyfill.io/v2/polyfill.min.js?features=IntersectionObserver"></script>
-  <script type="module" src="js/app.js"></script>
+  <script type="module" src="js/main.js"></script>
 </head>
 <body>
 <div id="player"></div>
