+++ WebApp/public/videoplayer/js/main.js
@@ -1,5 +1,5 @@
 import { VideoPlayer } from "./video-player.js";
-import { registerGamepadEvents, registerKeyboardEvents, registerMouseEvents, sendClickEvent } from "../../scripts/register-events.js";
+import { registerGamepadEvents, registerKeyboardEvents, registerMouseEvents, sendClickEvent } from "../../js/register-events.js";
 
 let playButton;
 let videoPlayer;
